making it responsive

+++ src/components/Adventures/Adventures.js
@@ -4,6 +4,7 @@ import { connect } from 'react-redux';
 import { getLocation, getDate, getTitle, getDetails } from './../../ducks/reducer'
 import axios from 'axios';
 import Header from '../Header/Header'
+import {Link} from 'react-router-dom'
 
 class Adventures extends Component {
     constructor(props) {
@@ -25,25 +26,43 @@ class Adventures extends Component {
     }
 
     deleteAdv(id) {
-        axios.delete('/api/adventures/' + id).then(resp => {
-            this.setState({
-                adventures: resp.data
+        var r = window.confirm('Are you sure you want to delete this Adventure?');
+        if (r === true){
+
+            axios.delete('/api/adventures/' + id).then(resp => {
+                this.setState({
+                    adventures: resp.data
+                })
             })
-        })
+        }
     }
 
     render() {
-                   
+        let { userData } = this.props;
         let newAdv = this.state.adventures.map((val, i) => {
             return (
-                <div className="Adventures_box">
-               
-                   <h2>{val.title}</h2>
-                    <h4>{val.user_name}</h4>
-                    <h6>{val.date}</h6>
-                    <div>Location: {val.location} </div>                   
-                    <p>{val.description}</p>
-                    <button className="Modal-delete-button" onClick={() => this.deleteAdv(val.id)}><img className="Delete-button" src="./delete.svg"/></button>
+                <div key={i} className="Adventures_box">
+                    {userData.user_name
+                        ?
+                        <div>
+                            <h2>{val.title}</h2>
+                            <h4>{val.user_name}</h4>
+                            <h6>{val.date}</h6>
+                            <div>Location: {val.location} </div>
+                            <p>{val.description}</p>
+                            <button className="Modal-delete-button" onClick={() => this.deleteAdv(val.id)}><img alt="" className="Delete-button" src="./delete.svg" /></button>
+                            <Link to={'edit/'+val.id}><img className="Edit-button" alt="" src="./edit.svg"/></Link>
+                        </div>
+                        :
+                        <div>
+                            <h2>{val.title}</h2>
+                            <h4>{val.user_name}</h4>
+                            <h6>{val.date}</h6>
+                            <div>Location: {val.location} </div>
+                            <p>{val.description}</p>
+                        </div>
+
+                    }
                 </div>
             )
         })
